Removing "re" in headers
The "re" command turned on the interpretation of the associated cell as a regex.  The usage vs. non-usage of "re" has proven a difficult point for the programmers, with programmers either forgetting "re" where it's necessary, or putting it in where it's not necessary.

We had "re" in the first place to try to maintain an early design decision, that all the content in cells was to be interpreted literally whenever possible, to avoid having to teach new programmers about special characters and escaping.  Instead, the programmer would "turn on" this specialness in some way -- basically, so that beginner programmers would not run into this snag before they knew about how to turn it on.  "re" was the command that turned this on.

In retrospect, this was probably not the right decision:

1. Programmers were not typically using special characters as literals in the first place.  (E.g., we expected that the programmers would be using lots of literals that might accidentally get interpreted as special, but mostly they're not.)  The most common special character programmers use, |, hasn't been used literally by anyone so far.

2. The opposite error -- e.g. the programmer *wanting* to use | specially and it's interpreted literally because they forgot use "re" -- is much more common.

This commit removes "re".  Regexes are now turned on by default under "equals", "starts", "ends", "contains", "from", "to", "pre", and "post", and there's no way to turn this off other than just escaping.  In the other direction, there's not currently any way to get a regex under normal headers like "text", you're just stuck to literals there now.  (But this may not be a bad thing.  Giving the programmer access to regexes outside of filters/rules is a recipe for unexpected and infinite results, and we don't currently have a clear use case for this in anyone's projects.)

This is a fairly far-reaching syntactic/semantic change, and so it introduces/uncovers a lot of bugs and breaks a lot of tests.  I haven't yet gotten around to all of them, but wanted to commit this as an intermediate checkpoint before things got too far from the original.

This uncovered an important bug that it turns out we weren't testing for: the behavior of empty regexes under equals was incorrect.  (It's uncovered now because *every* cell under an equals is now a regex, and some of those were empty.)

* Trivial regexes ("") generated an epsilon, but this wasn't actually correct if we want regexes to be used as filters.  (E.g. equals text:"" would select for an empty text field, whereas equals re text:"" wouldn't filter at all, it would let everything through.)  Now they generate an empty string on the regex dummy tape (.T).

* Erroneous regexes, though, should still produce epsilons.  We still want to default to letting everything through when there's a programmer error.  (Programmer errors that result in 0 outputs are very hard to catch/debug, so whenever possible we want to "fix" erroneous things in the direction of there being outputs rather than no outputs.)