Continuing to refactor to eliminate "re"
As mentioned in the last commit, there was still more to be done to eliminate "re".

(1) In the last commit, I narrowed what kinds of child headers from/to/pre/post/starts/ends/contains/equals could have.  After all, if they only accept plain tape names, why be broad about it?  But now I remembered why we did that, because even if a header is wrong, if you can parse it into *something*, you can give better error messages.

So I restored the broadness, now check for erroneous headers in the appropriate place, and added some new tests to testHeader.ts.  (Note that passing a testHeader test doesn't necessarily mean that header is valid, just that we can parse it.)

(2) In the last commit, I had temporarily separated out the regex-taking tag headers as individual objects (rather than generic TagHeaders) but this duplicated a lot of code; now they're back together as RegexTagHeaders.

(3) To give better error messages, I needed headers to know what they're called, this property is now .name.  There's a lot of conceptual overlap with .id here, so I simplified .id to mostly recapitulate the name.

(4) A strange bug arose, where starts/ends/contains started being concatenated with the previous grammar, rather than filtering it.  It turns out that in rescopeLeftBinders.ts, we weren't correctly testing whether the header was an instance of StartsHeader/EndsHeader/ContainsHeader etc.  (We should have been testing child.header.header, which is a Header, but we were only testing child.header, which is a TstHeader, that's in a different hierarchy.)  Anyway, this was an easy fix... but this code is quite old.  How was this ever working, how did things just break now?  In any case, it's working now.