Turning on "|" in plaintext content cells
There was one place where programmers *were* using "re" frequently outside of a filter, and that's to simplify tables where a morpheme can be, say, CLASS1 or CLASS2.  With the "re" command, this was easy: "re class: CLASS1|CLASS2".  Without the "re" command, expressing this means having two rows of the same morpheme, with the only difference being their "class" column.

Obviously, that's bad, that's some substantial repetition and therefore error.  So with "re" gone, we decided to make "|" available even in plaintext content cells, without any special command.  You can now just say "class: CLASS1|CLASS2".

This is the first character that's a "special character" in plain text.  We long said we wouldn't have these, but in the end making that true was causing programmer errors, and people weren't really using "|" anyway.  (There are a few languages that use that as part of their alphabet, but note that there's an alternative Unicode character specifically for that, U+01C0 LATIN LETTER DENTAL CLICK.)