Being explicit about spaces
For a long time we were cavalier about spaces in content cells.  Leading and trailing spaces of cells were stripped (since these are invisible to a spreadsheet user), but spaces inside strings were kept, they were just treated as characters.  But this was never actually standardized, it was just "whatever the parser/engine does with spaces, that's what happens".  We didn't even have any tests for this behavior, lest it change.

Now that we're parsing content cells, that forces us to be deliberate and explicit about what spaces mean and what should happen to them, and test for correct behavior.

In this commit, I make spaces work the same as in other parsed cells: that they're not themselves meaningful units (e.g. they don't represent a space literal), but can potentially serve to delimit/separate other objects.  This is the easiest thing to do (given our existing parsing engine) but also I think it's most consistent with our design philosophy and current usage.

    * E.g. consider "1SG | 2SG".  This should probably have the same meaning as "1SG|2SG". I think would be very surprising here if it actually meant the alternation of "1SG " and " 2SG", especially because those strings are otherwise inexpressible in the interface due to trimming.

    * But then this brings up the question of what "1SG 2SG" means.  If we consider this to be a literal space, then the programmer has to learn an algorithm for exactly when spaces are literal vs. delimiters, and this actually isn't trivial, it's a little tricky in places.

    * Also, programmers were already using the convention that _ meant space; no one was trying to represent literal spaces with spaces anyway.

So I think it's easiest just to say that spaces aren't literal.  They serve to potentially delimit elements but aren't themselves meaningful in any other way.  (NB: It's not *quite* accurate to say spaces are stripped out, it happens that content cells will always have the same outputs as their stripped equivalents, but that won't always be true of regexes.)

What if you need a literal space?  This commit also makes that possible by backslashing, like "1SG\ 2SG".  I think there are still some valid use cases for being able to *represent* a space, having *no* way to do this limits the language's ability to interact with real inputs/outputs.  (I don't think this should be the conventional way of representing spaces, I'm happy with _ for that, I just think that it should be *possible*.)

Because it's a change in how tokenization works, backslashing spaces is now also available in regexes, headers, and ops.  Note that there are still no valid uses of literal spaces in headers and ops, because spaces still aren't valid characters in identifiers.  It's just that the interpretation of spaces is unified between cell types, space has the same meaning everywhere and backslashed space has the same meaning everywhere.