Adjusting plaintext content parsing
(1) Because sequences and alternations had different scopes w.r.t. each other in plaintext content and regexes, there were some strings that would parse to a very different semantics.  Like "1SG 2SG|3SG" would have been "1SG 2SG" or "3SG" as plaintext, but "1SG 2SG" or "1SG 3SG" as a regex.  In the last commit I kind of bit the bullet on this, but really these should be identical.  (There are lots of places plaintext content and regexes differ, because they have different operators, interpret parens differently, etc., but this is a case where they share all these operators, so the interpretation should be the same.)  This commit reworks the plaintext grammar so that these have the same scope.

(2) Plaintext allowed "|1SG" to parse (as an empty string or "1SG"), but we decided some time ago that this should be an error, because it's actually quite easy, when adding and removing things to an alternation by copy and paste, to end up with things like "1SG|2SG|" or "1SG||2SG".  It's not obvious that there's an empty string there, and a more beginner programmer will probably not catch this.  This commit fixes this.

(3) A commit or two ago, we made MiniParseEnv distinguish separators from reserved words, and then MPUnreserved had a requirement that for something to be unreserved it had to be neither.  This isn't actually true in the big picture, something can potentially be a splitter without being reserved.  So now MPUnreserved only cares about MiniParseEnv.reserved, so make sure if a splitter is reserved it's also in .reserved.

(4) Added plaintext content tests.

(5) Started a grammar for symbol parsing, so that "Verb.Intransitive" can be parsed in a principled way rather than ad-hoc, because soon we'll get to the point where we'll need this.