Symbol name parsing
(2) The symbol name grammar committed last time didn't actually work, now it does.  It's a totally inefficient grammar, btw, but I think in practice it's not going to matter very much with how short the strings are that we're typically parsing; refactoring this for efficiency is way down our list of concerns.

(2) In the plaintext grammar, the toplevel expression was MPEmpty | MPRepetition(1, Infinity); this is equivalent to MPRepetition(0, Infinity).  (At one point there was a good reason for the repetition and emptiness to be different grammars, but some refactoring eliminated this reason.)

(3) Added tests for symbol parsing.

(4) Added Unicode tests for plaintext, symbol, and regex parsing.

(5) Fixed one of the Unicode tests for op parsing -- it was missing a ":" and so couldn't parse regardless of what it was testing for.