Sheet tests for valid/invalid IDs
We had a few of these, this commit adds a few more.

Also forgot to mention in the last commit message, all the reserved word globals are moved to a source file reserved.ts.  (Even though a recent commit made it so they shared the same list of reserved words, they still kept those lists independently.  Now that regex.ts uses that same list of reserved words for symbol parsing, it was time to unify them all.)