Header.toGrammar now a Pass
In the last commit, Regex.toGrammar was made a Pass; now Header.toGrammar is, too.

How is this helpful?  Well, for one, we're about to integrate the results of symbol parsing, which are also Regex objects, but with a different interpretation -- they don't become Grammars, at least not directly, they become alternations of symbol names.  I suppose I could have just added a .toSymbols() method to the Regex hierarchy, but at the point where we're doing potentially different things to the same hierarchy, it's nice to factor those out into Passes rather than complicate the Regex hierarchy with the different ways Regex objects can be interpreted.

(Come to think of it, we should rename those objects -- they're interpreted in more ways than just Regexes now.)