Various improvVarious improvements involving Count and CountTape.
1. This commit started off with changing unit tests to use CountTape instead
of Count. With this commit completed, the use of Count{Expr,Grammar} is limited
to use by the Interpreter to place an overall limit to avoid potentially
generating forever (although we often blow the stack before we hit the 1000
character limit). The only tests using Count are a couple in testInfinite that
exist to explicitly test the Count.potentiallyInfinite functionality, which is
important for the Interpreter to use CountGrammar.

2. An improvement was made to CountGrammar.potentiallyInfinite() and
CountTapeGrammar.potentiallyInfinite() to take into account whether the child
grammar is finite when maxChars == Infinite. Also, the case where only some
tapes are limited by a CountTapeGrammar was not accounted for in
CountTapeGrammar.potentiallyInfinite(); this has been fixed. Several tests have
been added to testInfinite to test CountTapeGrammar situations where specific
tapes are limited.

3. The above-mentioned changes to CountTapeGrammar.potentiallyInfinite()
revealed an issue in Interpreter.prepareExpr(), where for the case of a
PriorityGrammar inside a CollectionGrammar, the limiting CountGrammar was being
added at the start instead of after the PriorityGrammar, causing test 35a in
testReplace to fail where it previously passed. prepareExpr() has been fixed to
now insert the CountGrammar after the PriorityGrammar.

4. Some additional functions based around logging have been added to
testUtil. Specifically, determining the testSuiteName from the module has been
extracted out into a function called testSuiteName(). Also, logTestSuite() has
been added to output the name of the testSuite along with the date/time. A call
to this function has been added to the top in each test file updated in this
commit. (Others have been left for later.)

5. Finally, while mucking about in the test files for the first two items above,
various refactoring cleanup was done such as correcting, standardizing, and
numbering various test "describe" strings, extracting the expected results into
a variable instead of listing directly in the testGrammar call, formatting, etc.
ements involving Count and CountTape