Hooking up symbol parsing to embeds
Symbol names under embeds used to just be literal content.  Whatever the programmer wrote in that cell, we attempted to find it in the current scope and embed it.

In the last few commits, we've been working up to symbols being something that is actually parsed.  This is eventually for their inclusion into regexes like "q{HighVowel}.*", but this has some benefits to plain ol' embedded symbols as well:

  * It'll mean that symbol parsing under embeds and under other headers work the same way, leading to less drift/surprise.  (E.g., a particular way of phrasing a symbol might work under embed but not when inside a regex, which would be surprising.)  If they're parsed by the same grammar they won't be able to drift away from each other.

  * It lets the programmer say things like "embed: VerbIntrans|VerbTrans", which is a construction that the programmers actually use sometimes, and one that disappeared when we got rid of "re embed".

At the moment, doing this screws up a bit of our error messaging (because some things that used to just be "can't find this symbol" errors are now "can't parse this cell", which is a less helpful error message).  But in the long run, parsing at least gives us the *opportunity* for better error messages; the more you can parse a user input the more you can tell them where they got something wrong.