Symbols inside regexes
This now allows the embedding of symbols inside regexes with the use of curly braces, like "q{HighVowel}.*".

This doesn't make this *work* in the sheets syntax, there are still issues to resolve (in particular, about tape naming, this embedding has no idea about the ".T" dummy tape convention that regexes use).  All this does it let such regexes be parsed.

To make this work, we have to restrict the rules on valid identifiers a bit further.  Previously, there were some regex operators that were allowed inside identifiers, like "?", "+", and "*".

This isn't actually *ambiguous*, per se, but allowing identifiers to contain them when they're *also* tokenization separators in cells where identifiers can occur is a mess.  It's possible to do, but it's a mess, and that's weird anyway from a readability standpoint.  (Like if you saw a regex that said "{HighVowel?}", you would probably think that meant an optional HighVowel.  I would!  But it doesn't, this would mean a symbol called "HighVowel?".  That's not to say that I'm making this mean optional HighVowel now, the regex operators other than | are still not valid in symbol contexts, all I'm saying is that letting symbols contain "?" when symbols can occur in regexes is confusing.)

While I'm at it, I also removed "=" from the list of allowed identifier symbols, because we'll need that for something else shortly.