Cleaning up details and tests
(1) What do we do if the programmer wants to use curly brackets around identifiers outside of regexes?  E.g., under an embed.  Or even inside another pair of curly brackets, like "{{nouns}|{verbs}}"?

There's no syntactic reason why not, it's not ambiguous, although I don't like it because it blurs the boundary between which things are regexes and which aren't.  Programmers getting plaintext and regex contexts confused is a perennial issue, after all, we want to keep there being clear boundaries.

So the current solution is "Parse it (there's no reason not to parse it), but issue a warning that this isn't a place where you should use curly brackets."

(2) Fixed up a few tests in headers and ops that became invalid due to changes in allowed identifiers.

(3) testRegex.ts, and regex.ts itself, were increasingly about things that aren't regexes.  Renamed these to testCell.ts and cells.ts, and also renamed the previously-renamed testSheetContent.ts to testSheetCell.ts.

For now, I'm keeping the "Regex" object hierarchy named that, even though the collection of objects it represents is broader now.  I'm not sure what direction that hierarchy is going to go in the future anyway, it might not survive in its current form.