Assignment now uses = rather than :
One common syntax error was just writing "VERB" as an assignment rather than "VERB:".  The colon just isn't very visible there.  This makes "=" the assignment operator instead.

The other useful thing here is that it makes clearer the error when the programmer misspells an operator, or use an operator as an identifier.  This could lead to some odd error messages that don't make clear the real error -- like the parser thinking something is an assignment but in the wrong syntactic configuration, whereas the real problem is just that an operator was misspelled.  Distinguishing them syntactically helps the parser avoid this kind of unhelpful error.