Refactoring regex/header creation
The co-creation of header and cell grammars was a little bit all over the place -- not terribly, but I was having trouble keeping it all in my head, particularly because some of it was due to the inheritance structure of the Header hierarchy.  (E.g. "starts" causes regex parsing because StartsHeader inherits from RegexHeader.)

Now there's just a function getParseClass() that, given a header, tells you whether the cell underneath should be "regex" or "symbol" or "plaintext" etc.  You then pass that as an argument to a parseCell function rather than use specialized parseRegex, parseSymbol, etc. functions.

Rather than do this as Header methods (which still scatters the logic over a long file), I did it in pattern-matching style to get all the logic on the same page.