More header refactoring
Another thing we were doing in the Header hierarchy was determining the font color for the cells beneath, but this is now 100% predictable from the parseClass.  So now there's just a convenience function that translates the parseClass into a color string like "#000000".

Factoring this out, there's nothing left for RegexHeader to do, and we can get rid of it.  That's good, I wasn't liking how regexness was infecting Header's inheritance hierarchy, that's not a great way to handle the idea that something takes a regex.  (I mean, that was reasonable back when "re" was an actual header.  But it's not now, RegexHeader was something we were only using internally.)