Regexes with embedded symbols
Previous commits got symbols inside regexes (e.g., `equals text: q{HighVowel}`) working, but didn't yet attach them to the grammars.  This commit starts that; it's still not completely working, but I wanted to commit to at least record this stage of it.

Previously, regex terminals (like literals and dots) were implemented as being on a dummy tape .T, and when it came time to attach them to the rest of the grammar, like for "equals text: q", we'd just wrap the whole thing in a RenameGrammar from .T to text.

This wouldn't work for "equals text: {HighVowel}", HighVowel is already a grammar; it doesn't have any .T, it has its own text field.

So we have to do something a little different.  In this commit, I don't do the Rename thing anymore, but have a temporary grammar called SingleTapeGrammar.  All this really represents is the interface between the grammar as a whole (which needs the regex to be on a single tape named, say, text) and the grammar represented by the regex (which may have multiple tapes).

We can't turn this immediately into a grammar because, at the time of its construction, we don't know what HighVowel refers to, nor what tapes it might have.  We have to do name resolution and an initial calculateTapes run first.  Then we go through and convert the terminals, renaming the .T literals to literals of the relevant tape, and hiding tapes of embedded symbols that aren't the relevant tape.  (We can't just throw them out or ignore them, because they might be joining/filtering.)

This hiding of extra tapes isn't working correctly yet, but the rest is (we can at least replicate the functionality that existed before we got rid of "re embed"), so I wanted to commit what I had so far.