Conditions now Joins instead of Filters
In the previous commit, I mentioned that embedding a multi-tape symbol inside a regex (like `equals text: q{HighVowel}`) wasn't yet working.  The fix is simple -- basically one line changed -- although it'll take some time to explain.

A long time ago, we used joins to express equals/starts/etc.  (Actually this may predate starts/ends/contains.)  Later, we decided that what the programmer really means by these is a filter, not just a join -- that EVERY tape in the 2nd argument matches, not just that there's a match on tapes that they share.  If there happens to be a tape in the 2nd that the 1st doesn't share, there's just no output, you've made a condition that no record in the 1st can match.  (NB: This goes against our general lean towards "Programmer errors shouldn't cause zero outputs because those are very difficult to debug", but I'm not sure we had actually formulated this idea back then.)

This rarely came up, though -- it wasn't *easy* to equals/starts/etc. with a multi-tape grammar and programmers weren't using this construction.  And with the recent move towards only allowing single-tape conditions here, it was doubly moot.

Just recently, though, it did rear its head.  Using a filter rather than a join had consequences to embedding multi-tape symbols inside a regex.  As mentioned in the last commit, it's necessary to hide (rather than, say, drop or ignore) any tapes in `HighVowel` that aren't `text`.  But these tapes are still around, they're just hidden.

You might think that if you take a grammar X on T1, T2, T3 and hide T2 and T3, then filter Y[X], only T1 will be relevant for this filter.  It *could* be like that, but it wasn't, the tape-hiding systems and the join/filter systems don't know about each other.  Filters will demand that Y have the material on the now-hidden .T2 and .T3, too, and if it doesn't, they have no output.  So the hidden tapes of `HighVowel`, filtering a Y that doesn't share those tapes, were causing zero outputs.

We could change the filter grammar/expr so that it doesn't take into account hidden tapes, but it was more straightforward to go back to saying these are joins, and then the hidden tapes of `HighVowel` don't cause any trouble because they're not shared.  (This is probably ultimately better, anyway.  When the programmer says `equals text: whatever` and the grammar they're attaching that to doesn't have `text`, it's not useful to give them no outputs.  We should be handling that in the error messaging system, but still giving them outputs.  Join can do that, filter doesn't.)

Should we change filtering anyway, to ignore hidden tapes?  I'm going to leave it as-is.  Now there's only one place it's used in actual sheet-accessible grammars, which is for unit tests, and the programmer can't unit test for hidden tapes.  (Unit tests with hidden tapes are not syntactically expressible.)

Also, this change really made the naming of things counter-intuitive.  (Or it was already getting counter-intuitive, and this made it worse.)  FilterExpr meant the Y[X] relationship, but FilterGrammar meant the X part of that, and the Y[X] relationship was an EqualsGrammar, because it was the grammar created by the header `equals`.  But now `equals` no longer makes one of those, it makes a JoinGrammar.  The only place EqualsGrammar was used was to make unit tests -- that is, filters.

Really, the culprit is just not having a good word for the [X] part, so I've made it so that that's consistently called a "condition", freeing up the name FilterGrammar again, so we can call EqualsGrammar that and call it a day.

  FilterGrammar -> ConditionGrammar
  EqualsGrammar -> FilterGrammar
  Equals (convenience function) -> Filter