Cleanup of type hints
(1) Ever since we merged the TstComponent (now Component) and Grammar hierarchies into one, so that both have mapChildren() methods, it's been annoying to have to type hint the results of Grammar.mapChildren() (which are CResults) back into GrammarResults.

I didn't fully realize until recently that you could actually narrow the type of an overridden method A<B> => C<D> to A<B'> => C<D'> so long as B' inherits from B and D' from D.  That's the situation Grammar.mapChildren() is in.

So overriding it to mapChildren(p: GrammarPass, env: PassEnv): GrammarResult works, and that gets rid of the need for `as Grammar` and `as GrammarResult` type hints everywhere.

(2) Added .err() and .warn() convenience methods to Op, it was the last of the hierarchies not to have these.

(3) We still had a few `str as Token` type hints left, but Token is now just an alias for string, so these weren't doing anything.