More minor cleanup
As of the previous commit, all of the components X that can be Result<X>, as well as Result itself, had msg/err/warn methods.  However, the signatures of these functions were all subtly different.  Result.msg() could take a ResultVoid argument but none of the others could.  Component.err/warn and Grammar.err/warn could take a pos argument but the others couldn't.

This commit unifies the signatures of these.