Distinguishing hidden from internal tapes
We have a few tapes, like the input/output tapes for rules, where we wanted to make sure that they were distinct from any programmer tapes.  (E.g., we don't want to name them just "input" and "output", because those are entirely reasonable things to name the tapes of a transducer.) To make sure they were distinct from any possible programmer tapes, we prefixed these with the hidden tape prefix "."; programmers are unable to name anything starting with a ".".

Now, however, it's become clear that hiding needs to have semantic effects.  Previously, it was basically just cosmetic; hidden tapes didn't show up in the UI but otherwise they had the semantics of normal tapes.  That's beginning to change, with the realization that sometimes operations like negation and join may need to be sensitive to hiding.

Internal tapes like rule input/output are NOT hidden in this new sense, not at the point they're being used.  (Sometimes they're hidden later, sometimes they're renamed to something different.)

So we need to make a distinction now between "name of a hidden tape" and "name of a tape that's not hidden, we just need it to be distinct from programmer tapes".  We keep "." for the former, and now use "$" for the latter.  (At one point, the programmer *could* have named things starting with "$", but not anymore.  We're now stricter about what characters can begin an identifier.)

Nothing broke by making this change, which is reassuring; it makes clear that these tapes really weren't "hidden" in the true sense of the word.