Added a Not(Hide(X)) test
This commit just adds one test, but the test illustrates a problem whose simplicity makes it difficult.

In order to express things like "starts text: ~{HighVowel}" (the failing tests in testSheetFilter), we ultimately need a negation that's not sensitive to content on hidden tapes.  (E.g. a negation that negates content on the text tape, but doesn't negate content on the hidden height tape.)  Or more broadly, a kind of negation that's relativized only to particular tapes and lets other ones be.

This sounds like it would be easy, but it's not... at least, not the way we've structured things so far.  You'd think we could just go "Okay, if it's the tape we're negating, do the negation thing, and if it's not the tape we're negating, just pass through, act like a semantically empty wrapper class."  And sure, let's do that, but there's a snag.

Take this grammar.  I'll represent "Negation w.r.t. tape T1" as "Not_T1" here, and "Hide w.r.t tape T2" as "Hide_t2".

    (1) Not_t1(t1(foo) + Hide_t2(t2(bar)))

What's the derivative with respect to t1:a?  That's the case where we've "falled off" the negation graph, where we're already guaranteed to succeed regardless of what comes next on t1. Our code as written generates this:

    (2a) Rep(Dot(t1))

But that's not correct, we can't throw out the second part with the hidden t2.  The actual answer we want is this:

  (2) Rep(Dot(t1)) + Hide_t2(t2(bar))

Okay, that's not a hard formula, but how do we actually construct that from (1)?!?  It's not as easy as it looks!

We would need a mechanism that could actually get rid of, or replace, that t1(foo), but we don't have one.  Delta doesn't do that, delta NULLs out if the content's not empty, and would cause the whole sequence to NULL in turn.  It's also not a simple "grab the t1(foo) node from the child... think of this in general, there might be content equivalent to {t1:foo} that's generated any number of ways.  Really, it's a third recursive derivative function beyond deriv and delta, it's the old "DROP" function we hypothesized long ago.

I am quite resistant to the idea that the best way to solve this is a third derivative function!  I have two other ideas, which I'll try in the next commits.