Experimental HideExpr
Experimented with a HideExpr that pre-executes hidden content like PreTapeExpr, before realizing that it ran into the same issue with the un-constructable derivative described in the previous commit.

In a meeting on Wednesday, we determined that the real solution to the not-hide probably should be compilation to an automaton, and we're not quite ready to do that yet.  So for now we've decided to limit symbol embedding in regexes to single-tape grammars.

This commit also undoes a change in the previous commit, where hidden tapes were removed from the relevant tapes for negation.  This didn't end up being the solution to the problem, so let's keep it the way it was for now.