Bug in regex embeds in rules
ReplaceGrammar.constructExpr at one point calls delta to figure out if the "from" material is nullable.  If this material refers to a symbol (e.g. if it's `from: {vowels}`), then it's possible that the referent isn't yet in the symbol table, and throws an exception.

So we can't actually do this check at this stage; it needs to happen later after the symbol table is full.  That's going to be a real headache, this isn't the part of the pipeline that's been converted to "passes", so it's not easy to add another operation after the constructExpr.

I don't think this is optional, either, like I don't think we can just say "You can't use things like `{vowels}` in rules, we just have to bite the bullet and fix this.