Bugfix for embedded symbols in new-style rules
Testing the problematic grammars from the last commit with the new-style rules (i.e. the ones we actually use in sheets grammars) revealed several little bugs.

(1) When programming JoinRuleGrammar (the grammar that joins the new-style rules to the existing grammar), I had said "Oh, we don't have to actually calculate the tapes of the from/to/etc. grammars, we already know they're $i and $o, and outside of this grammar those won't exist anyway".  Which is true (or *will* be true) in the big picture, but that doesn't mean we don't have to CALCULATE it, because calculateTapes has a side effect of putting something in ._tapes.

So .tapes were undefined for children of rules.  We would ordinarily figure this out immediately (because if ._tapes is undefined accessing .tapes throws an exception) but it happens, amazingly, that we never needed to check this property for any children of rules in between creating the JoinRuleGrammar and eliminating it!  They all happened to have a trivial tape structure, and so we never needed to.

But when we can put an EmbedGrammar inside a rule, that's no longer true.  What tapes an EmbedGrammar refers to isn't trivial... I mean, that's basically why we have a .tapes member in the first place.  By not calling calculateTapes on the rule children of JoinRuleGrammar, EmbedGrammars within that scope weren't getting their ._tapes properly calculating, which was causing an exception when we tried to access .tapes.

(2) When thinking about SingleTapeGrammars, I had been assuming that it'd just be the users responsibility to make sure the single tape of the embedded grammar was the relevant tape.  (E.g., if you had a grammar "vowel" with "text", and you used it in "starts surface: {vowel}", then, well, too bad, there's no tape "surface" in "vowel".)  This was probably a bad decision even for equals/starts/etc., but it's a complete impossibility with new-style rules.  In rule like "{vowel} -> b", the relevant tape is "$i"!  The programmer can't even name something "$i", by design.

So now, it's no longer an error to have the grammar's single tape and the relevant tape be different.  If they're different, we do the rename behind the scenes to make it work.