Partial fix of symbol embedding in rules
This fix only gets rid of some instances of the problem, but it's a prereq for the full solution anyway, and I wanted to get it into its own commit.

The big-picture problem: ReplaceGrammar.constructExpr needs to know whether the rule's from-content is nullable, in which case they need to do things a little differently.  But nullability is something we only know about for sure by execution (i.e., by calling delta).  Computing delta requires a different environment than constructExpr, so constructExpr needs to make an ad-hoc environment in order to call delta.

The bug was that ReplaceGrammar.constructExpr was calling delta with a new, empty symbol table.  This was fine when rules couldn't have symbols inside!  But now, rules can have symbols inside!  So obviously, if we pass in an empty symbol table, it'll attempt to find those symbols, find nothing, and throw an exception.  (This general problem has caused bugs in the past as well, when environment in constructExpr has had different opinions about vocabulary compared to the one that's actually used later.)

This commit passes in the existing symbol table.  Grammar.constructExpr does have access to this object, through its parameterization -- this is the very symbol table that constructExpr is building.  But this doesn't solve our more general problem: the symbol table isn't yet complete, so whether or not the required expression exists depends on whether we've gotten around to constructing it yet.

This is somewhat of a different problem than our usual pass-priority issues -- even if we extended everything into micro-passes (as suggested in a recent commit message) it doesn't wholly resolve this problem.  Delta is something that's only defined in terms of the actual primitives of the system -- what we call Expr -- but we need that operation to GET the primitives of the system (in ReplaceGrammar.constructExpr).  Even if we were to merge the Grammar and Expr hierarchies and move to micropasses here, it wouldn't get rid of this more fundamental priority problem.

Luckily, the grammars that are allowed to be inside rules are simple -- we already decided to not allow recursion inside of them.  Since recursion is disallowed, the grammars themselves can't get into a priority problem.  (That is to say, the embedded symbol can't reference the rule that it itself is inside, so we can't get an actual chicken-and-egg problem, not when we're considering these as individual objects.)  There will always exist an order of symbols where, if we call constructExpr in that order, we'll not run into trouble.  The question is just determining that order, which I leave to the next commit.